requirements.txt
Django==1.9.4
django-autofixture==0.12.0
django-bootstrap-form==3.2
ecdsa==0.13
lxml==3.5.0
paramiko==1.16.0
proxmoxer==0.2.3
psycopg2==2.6.1
pycrypto==2.6.1
pycurl==7.43.0
pyldap==2.4.25.1
requests==2.9.1
